deleted helper file and am creating database objects in test file
